requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
sqlalchemy>=2.0.27
alembic>=1.13.1
psycopg[binary,pool]>=3.1.18
psycopg2-binary>=2.9.9
pgvector>=0.2.5
pydantic>=2.8.0
pydantic-settings>=2.1.0
python-multipart==0.0.9
python-dotenv==1.0.1
jieba==0.42.1
rank-bm25==0.2.2
mistune==3.0.2
openpyxl==3.1.2
httpx==0.27.0
aiofiles==23.2.1

# Office 文件解析（Skills 系统依赖）
# PDF 处理
pypdf>=3.0.0
pdfplumber>=0.10.0

# Office 文档处理
python-docx>=0.8.11
python-pptx>=0.6.21

# 开发依赖
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
ruff==0.2.2
mypy==1.8.0
